move findChildren and getParent into their own dir
