+++ src/utils.ts
@@ -1,7 +1,6 @@
 // Core Node.js imports for executing shell commands and handling paths
-import { exec, execSync, type StdioOptions } from 'child_process';
+import { execSync, type StdioOptions } from 'child_process';
 import type { ArgumentsCamelCase, BuilderCallback, CommandModule } from 'yargs';
-import { join } from 'path';
 
 /**
  * Interface for command configuration used by the CLI
@@ -82,127 +81,6 @@ export function getPrNumber(branch: string): number | null {
  */
 export function getPrStatus(prNum: number): PRStatus {
   return execCommand(`gh pr view ${prNum} --json state --jq '.state'`) as PRStatus;
-}
-
-/**
- * Formats the output of git branch --contains into a Set of branch names
- */
-function formatBranchContains(containsOutput: string, branchName: string) {
-  return new Set(containsOutput.split('\n')
-    .map((branch) => branch.replace('*', '').trim())
-    .filter(branch => branch && branch !== branchName));
-}
-
-/**
- * Determines the parent branch of a given branch
- * Handles both regular parents and stale parents (tagged with stale-parent)
- */
-export function getParentBranch(branchName: string): Branch {
-  const parentCommit = execCommand(`git rev-parse ${branchName}^`);
-  const parentBranchName = execCommand(`git branch --points-at ${parentCommit}`).trim();
-
-  if (parentBranchName) {
-    return {
-      branchName: parentBranchName,
-      parent: null,
-      children: [],
-      orphaned: false,
-      stale: false,
-    };
-  }
-
-  // Check for stale parent tags
-  const staleTag = execCommand(`git tag --points-at ${parentCommit} | grep -E '^stale-parent--figbranch--.+$'`);
-  const staleParentBranch = staleTag.split('--figbranch--')[1];
-
-  return {
-    branchName: staleParentBranch,
-    parent: null,
-    children: [],
-    orphaned: false,
-    stale: true,
-  };
-}
-
-/**
- * Finds all child branches of a given parent branch
- * Handles both current children and orphaned children (via stale tags)
- */
-export function findChildren(parentBranchName: string): Branch[] {
-  // Find current direct children
-  const parentCommit = execCommand(`git rev-parse ${parentBranchName}`);
-  const possibleCurrentChildren = formatBranchContains(
-    execCommand(`git branch --contains ${parentCommit}`),
-    parentBranchName
-  );
-
-  const currentChildren = new Set<string>();
-  possibleCurrentChildren.forEach(child => {
-    const actualParentCommit = execCommand(`git rev-parse ${child}^`);
-    if (parentCommit === actualParentCommit) {
-      currentChildren.add(child);
-    }
-  });
-
-  // Find orphaned children through stale tags
-  const staleTags = execCommand(`git tag | grep -E '^stale-parent--figbranch--${parentBranchName}--figbranch--[0-9]+$'`).split('\n');
-  const orphanedChildren = new Set<string>();
-
-  staleTags.forEach(tag => {
-    const tagCommit = execCommand(`git rev-parse ${tag}`);
-    const children = formatBranchContains(
-      execCommand(`git branch --contains $(git rev-parse ${tag})`),
-      parentBranchName
-    );
-    children.forEach(child => {
-      const parentCommit = execCommand(`git rev-parse ${child}^`);
-      if (parentCommit === tagCommit) {
-        orphanedChildren.add(child);
-      }
-    });
-  });
-
-  // Build branch objects for both current and orphaned children
-  const children: Branch[] = [];
-  const parent: Branch = {
-    branchName: parentBranchName,
-    parent: null,
-    children,
-    orphaned: false,
-    stale: false,
-  };
-
-  const staleParent: Branch = {
-    branchName: parentBranchName,
-    parent: null,
-    children,
-    orphaned: false,
-    stale: true,
-  }
-
-  // Add current children
-  currentChildren.forEach(child => {
-    children.push({
-      branchName: child,
-      parent: parent,
-      children: [],
-      orphaned: false,
-      stale: false,
-    });
-  });
-
-  // Add orphaned children
-  orphanedChildren.forEach(child => {
-    children.push({
-      branchName: child,
-      parent: staleParent,
-      children: [],
-      orphaned: true,
-      stale: false,
-    });
-  });
-
-  return children;
 }
 
 /**
